+++ src/main/java/es/in2/wallet/application/service/impl/UserDataUseCaseServiceImpl.java
@@ -3,7 +3,8 @@ package es.in2.wallet.application.service.impl;
 import es.in2.wallet.application.port.BrokerService;
 import es.in2.wallet.application.port.VaultService;
 import es.in2.wallet.application.service.UserDataUseCaseService;
-import es.in2.wallet.domain.model.CredentialsBasicInfoWithExpirationDate;
+import es.in2.wallet.domain.exception.NoSuchVerifiableCredentialException;
+import es.in2.wallet.domain.model.CredentialsBasicInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -27,10 +28,10 @@ public class UserDataUseCaseServiceImpl implements UserDataUseCaseService {
      * @param userId    The unique identifier of the user whose VCs are to be retrieved.
      */
     @Override
-    public Mono<List<CredentialsBasicInfoWithExpirationDate>> getUserVCs(String processId, String userId) {
+    public Mono<List<CredentialsBasicInfo>> getUserVCs(String processId, String userId) {
         return brokerService.getEntityById(processId, userId).flatMap(optionalEntity -> optionalEntity.map(userDataService::getUserVCsInJson).orElseGet(() -> {
             log.error("User with ID {} has no entity or credentials yet.", userId);
-            return Mono.error(new RuntimeException("There's no credential available."));
+            return Mono.error(new NoSuchVerifiableCredentialException("There is no credential available"));
         })).doOnSuccess(list -> log.info("Retrieved VCs in JSON for userId: {}", userId)).onErrorResume(Mono::error);
     }