+++ src/main/java/es/in2/wallet/domain/exception/handler/GlobalExceptionHandler.java
@@ -86,4 +86,27 @@ public class GlobalExceptionHandler {
                 .path(path)
                 .build());
     }
+    @ExceptionHandler(NoSuchVerifiableCredentialException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ResponseBody
+    public Mono<GlobalErrorMessage> noSuchVerifiableCredentialException(NoSuchVerifiableCredentialException ex, ServerHttpRequest request) {
+        String path = String.valueOf(request.getPath());
+        return Mono.just(GlobalErrorMessage.builder()
+                .title("NoSuchVerifiableCredentialException")
+                .message(ex.getMessage())
+                .path(path)
+                .build());
+    }
+
+    @ExceptionHandler(InvalidPinException.class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    @ResponseBody
+    public Mono<GlobalErrorMessage> invalidCredentialsException(InvalidPinException ex, ServerHttpRequest request) {
+        String path = String.valueOf(request.getPath());
+        return Mono.just(GlobalErrorMessage.builder()
+                .title("InvalidCredentialsException")
+                .message(ex.getMessage())
+                .path(path)
+                .build());
+    }
 }