+++ src/main/java/es/in2/wallet/domain/service/AuthorizationResponseService.java
@@ -6,4 +6,5 @@ import reactor.core.publisher.Mono;
 
 public interface AuthorizationResponseService {
     Mono<String> buildAndPostAuthorizationResponseWithVerifiablePresentation(String processId, VcSelectorResponse vcSelectorResponse, String verifiablePresentation, String authorizationToken) throws JsonProcessingException;
+    Mono<Void> sendDomeAuthorizationResponse(String vpToken, VcSelectorResponse vcSelectorResponse);
 }