+++ src/main/java/es/in2/wallet/infrastructure/core/controller/VerifiableCredentialController.java
@@ -3,7 +3,7 @@ package es.in2.wallet.infrastructure.core.controller;
 
 import es.in2.wallet.infrastructure.core.config.SwaggerConfig;
 import es.in2.wallet.application.service.UserDataUseCaseService;
-import es.in2.wallet.domain.model.CredentialsBasicInfoWithExpirationDate;
+import es.in2.wallet.domain.model.CredentialsBasicInfo;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
@@ -37,7 +37,7 @@ public class VerifiableCredentialController {
     @ApiResponse(responseCode = "200", description = "Verifiable credentials retrieved successfully.")
     @ApiResponse(responseCode = "400", description = "Invalid request.")
     @ApiResponse(responseCode = "500", description = "Internal server error.")
-    public Mono<List<CredentialsBasicInfoWithExpirationDate>> getVerifiableCredentialList(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader) {
+    public Mono<List<CredentialsBasicInfo>> getVerifiableCredentialList(@RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader) {
         log.debug("VerifiableCredentialController.getVerifiableCredential()");
 
         String processId = UUID.randomUUID().toString();