+++ src/test/java/es/in2/wallet/api/facade/UserDataUseCaseServiceImplTest.java
@@ -4,7 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import es.in2.wallet.application.service.impl.UserDataUseCaseServiceImpl;
-import es.in2.wallet.domain.model.CredentialsBasicInfoWithExpirationDate;
+import es.in2.wallet.domain.exception.NoSuchVerifiableCredentialException;
+import es.in2.wallet.domain.model.CredentialsBasicInfo;
 import es.in2.wallet.domain.service.UserDataService;
 import es.in2.wallet.application.port.VaultService;
 import es.in2.wallet.application.port.BrokerService;
@@ -45,11 +46,11 @@ class UserDataUseCaseServiceImplTest {
         Optional<String> userEntity = Optional.of(userEntityString);
 
 
-        String json = "{\"id\":\"subjectId\"}";
+        String jsonSubject = "{\"id\":\"subjectId\"}";
         ObjectMapper objectMapper2 = new ObjectMapper();
-        JsonNode credentialSubject = objectMapper2.readTree(json);
+        JsonNode credentialSubject = objectMapper2.readTree(jsonSubject);
 
-        List<CredentialsBasicInfoWithExpirationDate> expectedCredentials = List.of(new CredentialsBasicInfoWithExpirationDate("id1", List.of("type"), credentialSubject, ZonedDateTime.now()));
+        List<CredentialsBasicInfo> expectedCredentials = List.of(new CredentialsBasicInfo("id1", List.of("type"), List.of("jwt_vc","cwt_vc"),credentialSubject, ZonedDateTime.now()));
 
         when(brokerService.getEntityById(processId, userId)).thenReturn(Mono.just(userEntity));
         when(userDataService.getUserVCsInJson(anyString())).thenReturn(Mono.just(expectedCredentials));
@@ -70,7 +71,7 @@ class UserDataUseCaseServiceImplTest {
         when(brokerService.getEntityById(processId, userId)).thenReturn(Mono.just(Optional.empty()));
 
         StepVerifier.create(userDataFacadeService.getUserVCs(processId, userId))
-                .expectError(RuntimeException.class)
+                .expectError(NoSuchVerifiableCredentialException.class)
                 .verify();
         verify(brokerService).getEntityById(processId, userId);
         verifyNoInteractions(userDataService);