+++ src/test/java/es/in2/wallet/api/service/UserDataUseCaseServiceImplTest.java
@@ -323,13 +323,24 @@ class UserDataUseCaseServiceImplTest {
         vcValue.put(CREDENTIAL_SUBJECT, new LinkedHashMap<>(Map.of("id", "subjectId")));
         vcValue.put(EXPIRATION_DATE, "2024-04-07T09:57:59Z");
 
-        VCAttribute vcAttribute = new VCAttribute("vcId", VC_JSON, vcValue);
+        List<VCAttribute> mockVcs = List.of(
+                VCAttribute.builder()
+                        .id("vc1")
+                        .type(VC_JSON)
+                        .value(vcValue)
+                        .build(),
+                VCAttribute.builder()
+                        .id("vc1")
+                        .type(JWT_VC)
+                        .value("ey24343...")
+                        .build()
+        );
 
         UserEntity userEntity = new UserEntity(
                 "user123",
                 "userEntity",
                 new EntityAttribute<>("Property", List.of()),
-                new EntityAttribute<>("Property", List.of(vcAttribute))
+                new EntityAttribute<>("Property", mockVcs)
         );
 
         when(objectMapper.readValue(anyString(), eq(UserEntity.class))).thenReturn(userEntity);
@@ -339,13 +350,52 @@ class UserDataUseCaseServiceImplTest {
         StepVerifier.create(userDataServiceImpl.getUserVCsInJson(userEntityString))
                 .assertNext(credentialsBasicInfoWithExpiredDate -> {
                     assertEquals(1, credentialsBasicInfoWithExpiredDate.size());
-                    CredentialsBasicInfoWithExpirationDate credentialsInfo = credentialsBasicInfoWithExpiredDate.get(0);
-                    assertEquals("vcId", credentialsInfo.id());
+                    CredentialsBasicInfo credentialsInfo = credentialsBasicInfoWithExpiredDate.get(0);
+                    assertEquals("vc1", credentialsInfo.id());
                     assertEquals(List.of("VerifiableCredential", "SpecificCredentialType"), credentialsInfo.vcType());
+                    assertEquals(List.of("jwt_vc"), credentialsInfo.availableFormats());
                     assertEquals("subjectId", credentialsInfo.credentialSubject().get("id").asText());
                     assertEquals(ZonedDateTime.parse("2024-04-07T09:57:59Z"), credentialsInfo.expirationDate());
                 })
                 .verifyComplete();
+    }
+
+    @Test
+    void testGetUserVCsInJsonDateTimeParseException() throws Exception {
+        String userEntityString = "userEntityJsonString";
+        String invalidDate = "2024434-04-07T09:57:59Z";
+        LinkedHashMap<String, Object> vcValue = new LinkedHashMap<>();
+        vcValue.put("type", List.of("VerifiableCredential", "SpecificCredentialType"));
+        vcValue.put(CREDENTIAL_SUBJECT, new LinkedHashMap<>(Map.of("id", "subjectId")));
+        vcValue.put(EXPIRATION_DATE, invalidDate);
+
+        List<VCAttribute> mockVcs = List.of(
+                VCAttribute.builder()
+                        .id("vc1")
+                        .type(VC_JSON)
+                        .value(vcValue)
+                        .build(),
+                VCAttribute.builder()
+                        .id("vc1")
+                        .type(JWT_VC)
+                        .value("ey24343...")
+                        .build()
+        );
+
+        UserEntity userEntity = new UserEntity(
+                "user123",
+                "userEntity",
+                new EntityAttribute<>("Property", List.of()),
+                new EntityAttribute<>("Property", mockVcs)
+        );
+
+        when(objectMapper.readValue(anyString(), eq(UserEntity.class))).thenReturn(userEntity);
+        when(objectMapper.convertValue(any(LinkedHashMap.class), eq(JsonNode.class)))
+                .thenReturn(new ObjectMapper().valueToTree(vcValue));
+
+        StepVerifier.create(userDataServiceImpl.getUserVCsInJson(userEntityString))
+                .expectError()
+                .verify();
     }
 
     @Test